测试

+++ src/main/java/com/xiliulou/electricity/entity/BatteryOtherProperties.java
@@ -85,45 +85,45 @@ public class BatteryOtherProperties {
     /**
      * 	gsm信号强度 01106001
      */
-    private Integer gsmSignalStrength;
+    private Long gsmSignalStrength;
     /**
      * 	电池状态 01108001
      */
-    private Integer batteryStatus;
+    private Long batteryStatus;
     /**
      * 	速度信息 01109001
      */
-    private Integer speed;
+    private Long speed;
     /**
      * 	行驶里程 01110001
      */
-    private Integer tripMiles;
+    private Long tripMiles;
     /**
      * 	总电压 01111001
      */
-    private Integer sumV;
+    private Long sumV;
     /**
      * 	总电流 01112001
      */
-    private Integer sumA;
+    private Long sumA;
 
     /**
      * 	功率温度值  01118001
      */
-    private Integer powerTemp;
+    private Long powerTemp;
 
     /**
      * 	总放电 01124001
      */
-    private Integer sumDischarge;
+    private Long sumDischarge;
     /**
      * 	总充电 01125001
      */
-    private Integer sumCharge;
+    private Long sumCharge;
     /**
      * 	预计放电时间 01126001
      */
-    private Integer expDischargeTime;
+    private Long expDischargeTime;
 
 
 }
